Move `ephemeral` package from `net` to `crypto`
The net/ephemeral package was moved out of the
net package, into the crypto/ephemeral package,
because the ECDH handshake/encryption does not
happen on the network level. It is one of the
steps of GJKR. Key types used inside the
ephemeral package do not depend on key types
used by libp2p.
