+++ pkg/beacon/relay/gjkr/marshaling_test.go
@@ -10,7 +10,7 @@ import (
 	bn256 "github.com/ethereum/go-ethereum/crypto/bn256/cloudflare"
 	"github.com/keep-network/keep-core/pkg/beacon/relay/group"
 	"github.com/keep-network/keep-core/pkg/internal/pbutils"
-	"github.com/keep-network/keep-core/pkg/net/ephemeral"
+	"github.com/keep-network/keep-core/pkg/crypto/ephemeral"
 )
 
 func TestEphemeralPublicKeyMessageRoundtrip(t *testing.T) {
