+++ pkg/beacon/relay/gjkr/member.go
@@ -5,7 +5,7 @@ import (
 
 	bn256 "github.com/ethereum/go-ethereum/crypto/bn256/cloudflare"
 	"github.com/keep-network/keep-core/pkg/beacon/relay/group"
-	"github.com/keep-network/keep-core/pkg/net/ephemeral"
+	"github.com/keep-network/keep-core/pkg/crypto/ephemeral"
 )
 
 type memberCore struct {
