+++ pkg/beacon/relay/gjkr/message_test.go
@@ -7,7 +7,7 @@ import (
 	"testing"
 
 	"github.com/keep-network/keep-core/pkg/beacon/relay/group"
-	"github.com/keep-network/keep-core/pkg/net/ephemeral"
+	"github.com/keep-network/keep-core/pkg/crypto/ephemeral"
 )
 
 func TestAddAndDecryptSharesFromMessage(t *testing.T) {
