+++ pkg/crypto/ephemeral/interface.go
