+++ pkg/crypto/ephemeral/symmetric_key.go
